Simplify http-classify stack initialization
Remove the concrete `Layer` type in favor of an anonymous helper and
rename service types to match our newer idioms.
