+++ linkerd/app/inbound/src/lib.rs
@@ -257,7 +257,7 @@ impl Config {
                     .push(metrics.http_route.into_layer::<classify::Response>())
                     // Sets the per-route response classifier as a request
                     // extension.
-                    .push(classify::Layer::new())
+                    .push(classify::NewClassify::layer())
                     .check_new_clone::<dst::Route>()
                     .push_map_target(endpoint::route)
                     .into_inner(),
